+++ REST/src/main/java/org/example/rest/services/RentService.java
@@ -2,9 +2,16 @@ package org.example.rest.services;
 
 import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoClient;
+import org.bson.types.ObjectId;
+import org.example.rest.exceptions.BadRequestException;
+import org.example.rest.exceptions.ConflictException;
 import org.example.rest.exceptions.NotFoundException;
+import org.example.rest.models.Book;
 import org.example.rest.models.Rent;
+import org.example.rest.models.User;
+import org.example.rest.repositories.BookRepository;
 import org.example.rest.repositories.RentRepository;
+import org.example.rest.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
@@ -14,14 +21,24 @@ import java.util.List;
 @Service
 public class RentService {
     private final RentRepository rentRepository;
+    private final UserRepository userRepository;
+    private final BookRepository bookRepository;
     private final MongoClient mongoClient;
 
-    public RentService(RentRepository rentRepository, @Qualifier("mongoClient") MongoClient mongoClient) {
+    public RentService(RentRepository rentRepository, UserRepository userRepository, BookRepository bookRepository, @Qualifier("mongoClient") MongoClient mongoClient) {
         this.rentRepository = rentRepository;
+        this.userRepository = userRepository;
+        this.bookRepository = bookRepository;
         this.mongoClient = mongoClient;
     }
 
     public Rent getRentById(String id) {
+        boolean idIsValid = ObjectId.isValid(id);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         Rent rent = rentRepository.findById(id);
 
         if (rent == null) {
@@ -36,18 +53,42 @@ public class RentService {
     }
 
     public List<Rent> getCurrentRentsByUserId(String userId) {
+        boolean idIsValid = ObjectId.isValid(userId);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         return rentRepository.findByUserIdAndEndDateIsNull(userId);
     }
 
     public List<Rent> getArchiveRentsByUserId(String userId) {
+        boolean idIsValid = ObjectId.isValid(userId);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         return rentRepository.findByUserIdAndEndDateIsNotNull(userId);
     }
 
     public List<Rent> getCurrentRentsByBookId(String bookId) {
+        boolean idIsValid = ObjectId.isValid(bookId);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         return rentRepository.findByBookIdAndEndDateIsNull(bookId);
     }
 
     public List<Rent> getArchiveRentsByBookId(String bookId) {
+        boolean idIsValid = ObjectId.isValid(bookId);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         return rentRepository.findByBookIdAndEndDateIsNotNull(bookId);
     }
 
@@ -55,16 +96,50 @@ public class RentService {
         try (ClientSession clientSession = mongoClient.startSession()) {
             clientSession.startTransaction();
 
+            User existingUser =  userRepository.findById(rent.getUserId());
+
+            if (existingUser == null) {
+                clientSession.abortTransaction();
+                throw new NotFoundException("Nie znaleziono użytkownika o podanym ID");
+            }
+
+            Book existingBook =  bookRepository.findById(rent.getBookId());
+
+            if (existingBook == null) {
+                clientSession.abortTransaction();
+                throw new NotFoundException("Nie znaleziono książki o podanym ID");
+            }
+
+            if (!existingUser.isActive()) {
+                clientSession.abortTransaction();
+                throw new ConflictException("Użytkownik nie jest aktywny");
+            }
+
+            List<Rent> currentRents = rentRepository.findByBookIdAndEndDateIsNull(existingBook.getId());
+
+            if (!currentRents.isEmpty()) {
+                clientSession.abortTransaction();
+                throw new ConflictException("Książka jest już wypożyczona.");
+            }
+
             Rent createdRent = rentRepository.create(rent);;
 
             clientSession.commitTransaction();
             return createdRent;
+        } catch (NotFoundException | ConflictException e) {
+            throw e;
         } catch (RuntimeException e) {
             throw new RuntimeException("Wystąpił błąd podczas tworzenia wypożyczenia.");
         }
     }
 
     public Rent endRent(String id) {
+        boolean idIsValid = ObjectId.isValid(id);
+
+        if (!idIsValid) {
+            throw new BadRequestException("To nie jest prawidłowy format id, powinien mieć 24 znaki w zapisane w hex");
+        }
+
         try (ClientSession clientSession = mongoClient.startSession()) {
             clientSession.startTransaction();
 
@@ -75,17 +150,20 @@ public class RentService {
                 throw new NotFoundException("Nie znaleziono wypożyczenia o podanym ID");
             }
 
-            existingRent.setEndDate(LocalDateTime.now());
+            LocalDateTime now = LocalDateTime.now();
+
+            if (now.isBefore(existingRent.getBeginDate())) {
+                throw new ConflictException("Data zakończenia nie może być wcześniejsza niż data rozpoczęcia wypożyczenia");
+            }
+
+            existingRent.setEndDate(now);
 
             Rent updatedRent = rentRepository.update(existingRent);
 
-            if (updatedRent == null) {
-                clientSession.abortTransaction();
-                throw new RuntimeException("Nie udało się zaktualizować wypożyczenia.");
-            }
-
             clientSession.commitTransaction();
             return updatedRent;
+        } catch (NotFoundException | ConflictException e) {
+            throw e;
         } catch (RuntimeException e) {
             throw new RuntimeException("Wystąpił błąd podczas aktualizacji wypożyczenia.");
         }
@@ -101,16 +179,18 @@ public class RentService {
                 throw new NotFoundException("Nie znaleziono wypożyczenia o podanym ID");
             }
 
-            boolean deleteSuccess = rentRepository.delete(id);
-
-            if (!deleteSuccess) {
+            if (existingRent.getEndDate() != null) {
                 clientSession.abortTransaction();
-                throw new RuntimeException("Nie udało się usunąć książki.");
+                throw new ConflictException("Nie można usunąć zakończonego wypożyczenia.");
             }
 
+            rentRepository.delete(id);
+
             clientSession.commitTransaction();
+        } catch (NotFoundException | ConflictException e) {
+            throw e;
         } catch (RuntimeException e) {
-            throw new RuntimeException("Wystąpił błąd podczas aktualizacji książki.");
+            throw new RuntimeException("Wystąpił błąd podczas usuwania wypożyczenia.");
         }
     }
 }
